Update Data for Unicode 8
This change updates:
* Unihan_NumericValues.txt
* Unihan_OtherMappings.txt
* CJKRadicals.txt
* Unihan_Variants.txt
* Unihan_DictionaryIndices.txt
* Unihan_DictionaryLikeData.txt

to the latest available versions, created on May 1st 2015, for Unicode 8.

This first set of files work without any changes to the existing code,
and bring the following improvements for Pin1yin1:
* improved stroke count (kTotalStrokes was redefined)
* Updated range of covered characters to include U+9FCD through U+9FD5
* Added references to CJK Unified Ideographs Extension E block.

The remaining files in the Unicode 8 update:
Unihan_IRGSources.txt
Unihan_RadicalStrokeCounts.txt
Unihan_Readings.txt

will require code changes, since the format of pinyin records changes
from using ascii with a number to using accents. In addition, certain
parts of these files, such as the kRSUnicode are moved to a new file:
Unihan_IRGSources.txt . As such, these will come in a future patch.