Update Readings data for Unicode 8
This change updates:
* Unihan_Readings.txt

and is the final piece of data to update Pin1yin1 to Unicode 8

From the previous data to Unicode 8, the kMandarin field was redfined.
Particularly for this application:
* the data now uses accent marks rather than tone numbers
* only primary pronunciations are listed - no more secondary pronunciations.

This required some small updates to the import_cedict task, to convert
the Unicode data to ascii for comparison with the cedict data, and
simplification of the primary/secondary pronunciations code.