Fix special character encoding in pinyin.js
EncodeURI does not encode special characters: "; , / ? : @ & = + $",
which meant that various combinations of these - particularly the ampersand,
could prevent or truncate conversions.

This patch replaces EncodeURI with EncodeURIComponent, which is the same
except that it converts those special characters too.

This works for the context in which it is used - the part of the URL
following the hash.

fixes #17